print-api/requirements.txt
flask>=3.0
python-escpos>=3.0
python-dotenv>=1.0
Pillow>=10.0
Jinja2>=3.1
flask-limiter>=3.5
